requirements.txt
fastapi
uvicorn
google-cloud-storage
duckdb
pillow
numpy
jinja2
python-multipart

#HOG features, skimage
scikit-image
matplotlib

#ai classifier, models
transformers
torch

#histogram 
opencv-python-headless==4.10.0.84
